+++ packages/_private/e2e-demo-test/src/DemoFormValidate.spec.ts
@@ -38,17 +38,6 @@ createTestCases(({ name, port }) => {
       const form = new DemoFormValidate(page, port);
       await form.loadInitialPage();
 
-      // react on submit / confirm dialog
-      let submitted = false;
-      page.on('dialog', dialog => {
-        submitted = true;
-        dialog
-          .accept()
-          .catch(() => {
-            submitted = false;
-          });
-      });
-
       // check initial state
       await form.checkInitialState(expect);
 
@@ -58,9 +47,14 @@ createTestCases(({ name, port }) => {
       // submit valid form
       await form.submit.click();
 
-      expect(submitted).toBe(true);
+      await expect(await form.form.evaluate(
+        (f) => (f as HTMLFormElement).checkValidity(),
+      )).toBe(true);
 
-      expect(await form.form.evaluate((f) => (f as HTMLFormElement).checkValidity())).toBe(true);
+      // Check that the success message is shown
+      await expect(form.formStatus).toHaveAttribute('open');
+      await expect(form.formStatus).toHaveAttribute('variant', 'success');
+      await expect(form.formStatus).toContainText('successfully');
     });
   });
 });