+++ packages/_private/e2e-demo-test/src/PageObjects/DemoFormValidate.ts
@@ -26,30 +26,51 @@ export class DemoFormValidate extends BaseFormObject {
   }
 
   async fill() {
+    // Custom timeout.
+    // This IS needed for react 19.2.0 to function properly with playwright
+    // You may tweak this to lower numbers if it really matters,
+    // but 20ms is a good balance and was not flaky in tests
+    const timeout = 20;
+
     await this.gender.getByText('Female').check();
+    await this.page.waitForTimeout(timeout);
+
     await fillField(this.name, 'Maxim');
+    await this.page.waitForTimeout(timeout);
+
     await fillField(this.email, 'max@musterman.de');
+    await this.page.waitForTimeout(timeout);
 
     await this.role.click();
+    await this.page.waitForTimeout(timeout);
+
     await this.frontend.click();
+    await this.page.waitForTimeout(timeout);
 
     await fillField(this.password, 'Password123');
+    await this.page.waitForTimeout(timeout);
+
     await fillField(this.nationality, 'German', '.combobox__display-input');
+    await this.page.waitForTimeout(timeout);
 
     await this.newsSyn.click({
       position: { x: 20, y: 20 },
     });
+    await this.page.waitForTimeout(timeout);
 
     await this.newsBeta.click({
       position: { x: 20, y: 20 },
     });
+    await this.page.waitForTimeout(timeout);
 
     await fillTextArea(this.additionalInfo, 'Comment');
+    await this.page.waitForTimeout(timeout);
 
     // Drag the happiness handle to a 9 of 10 rating
     await this.happiness.dragTo(this.happiness, {
       targetPosition: { x: 1000, y: 50 },
     });
+    await this.page.waitForTimeout(timeout);
   }
 
   async checkInitialState(expect: Expect) {